Moved DeviceService out of E4Service and abstracted MainActivity
