+++ app/src/main/java/org/radarcns/android/DeviceServiceBinder.java
@@ -1,5 +1,7 @@
 package org.radarcns.android;
 
+import android.support.annotation.NonNull;
+
 import org.apache.avro.specific.SpecificRecord;
 import org.radarcns.data.Record;
 import org.radarcns.kafka.AvroTopic;
@@ -7,13 +9,17 @@ import org.radarcns.kafka.rest.ServerStatusListener;
 import org.radarcns.key.MeasurementKey;
 
 import java.util.List;
+import java.util.Set;
 
 public interface DeviceServiceBinder {
-    /** Start scanning and recording from a compatible device. */
-    DeviceState startRecording();
+    /** Start scanning and recording from a compatible device.
+     * @param acceptableIds a set of source IDs that may be connected to.
+     *                      If empty, no selection is made.
+     */
+    DeviceState startRecording(@NonNull Set<String> acceptableIds);
     /** Stop scanning and recording */
     void stopRecording();
-    <V extends SpecificRecord> List<Record<MeasurementKey, V>> getRecords(AvroTopic<MeasurementKey, V> topic, int limit);
+    <V extends SpecificRecord> List<Record<MeasurementKey, V>> getRecords(@NonNull AvroTopic<MeasurementKey, V> topic, int limit);
     /** Get the current device status */
     DeviceState getDeviceStatus();
     /** Get the current server status */
